git-svn-id: file:///home/danlipsa/dan-svn-repo/src@2227 d596a3c1-de11-4375-84b4-10211c464620

+++ ViewSettings.h
@@ -270,7 +270,7 @@ public:
     {
 	m_angleOfView = value;
     }
-    double GetCameraDistance () const
+    float GetCameraDistance () const
     {
 	return m_cameraDistance;
     }
@@ -579,7 +579,7 @@ private:
     /**
      * Distance from the camera to the center of the bounding box for the foam.
      */
-    double m_cameraDistance;
+    float m_cameraDistance;
 
     //Stationary
     AverageAroundType m_averageAroundType;
